requirements.txt
Django==5.0.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
django-autoslug==1.9.9
django-environ==0.11.2
django-minio-backend==3.5.0
django-restql==0.15.4
djangorestframework==3.14.0
minio==7.2.5
psycopg2-binary==2.9.9
pip==24.0
pyPEG2==2.15.2
pycparser==2.21
pycryptodome==3.20.0
pytz==2024.1
Pillow
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1